# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for Golang Project Schedule Manager.",
        "title": "Schedule Manager Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "andreipetrov080@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/workspace": {
            "put": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Update workspace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspace"
                ],
                "parameters": [
                    {
                        "description": "Workspace data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/workspace.WorkspaceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.WorkspaceData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Create workspace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspace"
                ],
                "parameters": [
                    {
                        "description": "Workspace data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/workspace.WorkspaceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.WorkspaceData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/workspace/task/{id}": {
            "delete": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Remove workspace by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspace"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace task id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/success.ServerSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspace/{id}": {
            "get": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Get workspace by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspace"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.WorkspaceData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Remove workspace by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspace"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/success.ServerSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspace/{id}/invite": {
            "get": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Get invite token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspace"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/workspace.WorkspaceInviteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspace/{id}/member": {
            "post": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Create workspace task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspace"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Workspace member data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/workspace.WorkspaceMemberRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.WorkspaceMember"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/workspace/{id}/task": {
            "post": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Create workspace task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspace"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Workspace task data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/workspace.WorkspaceTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.WorkspaceTask"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/me": {
            "get": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Get current user workspaces",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspace"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.WorkspaceData"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/search": {
            "get": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Search workspaces",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspace"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace name",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.WorkspaceData"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.ServerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.Workspace": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "domain.WorkspaceData": {
            "type": "object",
            "required": [
                "workspace"
            ],
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.WorkspaceMember"
                    }
                },
                "queues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.WorkspaceQueue"
                    }
                },
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.WorkspaceSchedule"
                    }
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.WorkspaceTask"
                    }
                },
                "workspace": {
                    "$ref": "#/definitions/domain.Workspace"
                }
            }
        },
        "domain.WorkspaceMember": {
            "type": "object",
            "required": [
                "member",
                "status",
                "workspaceId"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "member": {
                    "$ref": "#/definitions/domain.User"
                },
                "status": {
                    "type": "string"
                },
                "workspaceId": {
                    "type": "integer"
                }
            }
        },
        "domain.WorkspaceQueue": {
            "type": "object",
            "required": [
                "name",
                "workspaceId"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string"
                },
                "workspaceId": {
                    "type": "integer"
                }
            }
        },
        "domain.WorkspaceSchedule": {
            "type": "object",
            "required": [
                "records",
                "startDate",
                "workspaceId"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.WorkspaceScheduleRecord"
                    }
                },
                "startDate": {
                    "type": "string"
                },
                "workspaceId": {
                    "type": "integer"
                }
            }
        },
...
        "domain.WorkspaceTask": {
            "type": "object",
            "required": [
                "name",
                "workspaceId"
            ],
            "properties": {
                "creator": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "workspaceId": {
                    "type": "integer"
                }
            }
        },
        "error.ServerErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "success.ServerSuccessResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "workspace.WorkspaceInviteResponse": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "workspace.WorkspaceMemberRequest": {
            "type": "object",
            "required": [
                "email",
                "status"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "workspace.WorkspaceRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "workspace.WorkspaceTaskRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AuthorizationKey": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}